requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==6.0.0
PyJWT==2.8.0
PyMySQL==1.1.1
Werkzeug==3.1.4
python-dotenv==1.0.0
jieba==0.42.1
numpy==2.1.3
opencv-python==4.10.0.84
scikit-learn==1.6.1
scikit-image==0.25.0
Pillow==11.1.0
requests==2.31.0
openai==1.12.0
